+++ backend/curator/src/main/java/com/ssafy/curator/repository/post/CommentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
-    List<CommentEntity> findByPostId(int id);
-    CommentEntity findById(int id);
+    List<CommentEntity> findByPostId(Long id);
+    CommentEntity findById(Long id);
 }
